Revamping TSP Tests to Pytest
